Fixed clean invalid block migrations

+++ admin/migrations/clean-invalid-blocks-migration.php
@@ -20,13 +20,15 @@ class Brizy_Admin_Migrations_CleanInvalidBlocksMigration implements Brizy_Admin_
 			global $wpdb;
 
 			$invalidBlocks = $wpdb->get_results(
-				"SELECT ID FROM wp_posts p 
-						LEFT JOIN wp_postmeta m ON m.post_id=p.id and m.meta_key='brizy'
+				"SELECT ID FROM {$wpdb->posts} p 
+						LEFT JOIN {$wpdb->postmeta} m ON m.post_id=p.id and m.meta_key='brizy'
 						WHERE p.post_type='brizy-global-block' and m.meta_value is NULL" );
 
-			foreach ( $invalidBlocks as $block ) {
-				$wpdb->query( "DELETE FROM {$wpdb->posts} WHERE ID={$block->ID}" );
-				$wpdb->query( "DELETE FROM {$wpdb->postmeta} WHERE post_id={$block->ID}" );
+			if ( is_array( $invalidBlocks ) ) {
+				foreach ( $invalidBlocks as $block ) {
+					$wpdb->query( "DELETE FROM {$wpdb->posts} WHERE ID={$block->ID}" );
+					$wpdb->query( "DELETE FROM {$wpdb->postmeta} WHERE post_id={$block->ID}" );
+				}
 			}
 
 		} catch ( Exception $e ) {
